DEBT: update README and documentation
