+++ README.md
@@ -2,17 +2,80 @@
 
 [![Go Doc](https://img.shields.io/badge/godoc-reference-blue.svg?style=flat-square)](https://github.com/super-type/supertype)
 
-This is where we will keep our backend functionality in the form of a smart monolith, for the time being. The microservices-first path is wrought with peril, and we wish to make our lives as easy as possible. Which, at a startup, is a challenge in and of itself. Many of the best minds in the software architecture space ([Martin Fowler](https://martinfowler.com/bliki/MonolithFirst.html), [Atlassian Blog](https://www.atlassian.com/continuous-delivery/microservices/building-microservices)) argue that starting with a monolith is the simplest way to get code off the ground.
+This repo is the backbone of Supertype's device-agnostic data network. Almost all backend functionality lives within this repo, as our service is not yet large enough to justify splitting into multiple repositories. This monorepo acts as the central server for our system, and all data ingress and egress moves through the HTTP rotuer. Current contexts supported are:
 
-Building an architecture is inherently complicated, and we're inherently bad at that. As a startup, requirements, needs, and client requests are constantly changing, and we want to start with an architecture that's best suited for that. While we have put in a signficant amount of effort into determing what our end-state, microservices architecture may look like, we don't have a crystal ball.
+* Authenticating: How vendors and users register and authenticate and authorize communication with Supertype.
+* Consuming: How vendors can receive data from Supertype's data network.
+* Producing: How vendors can produce data to Supertype's data netowrk.
+* Vendor Management: Fetching vendor information regarding account infromation, payments, etc - primarily used for billing and vendor dashboard purposes
+* Attribute Management: The constantly-solidifying mechanism of how vendors produce and consume data to and from the Supertype data network.
 
-Therefore, we will start with a monolith. Once we have a specific service within our monolith that warrants a microservices, we'll spin it out to become our first microservice. Until then, monolith it is.
+## Terms and definitions
+**vendor:** a smart home device maker. These are the companies actually using Supertype.
 
-## Architecture
+**user:** a vendor's end user. The individual smart home owner/operator who is using vendor devices.
 
-![Architecture](internal/images/architecture-9-10.png?raw=true "Architecture")
+**attribute:** a specific entity that vendor devices observe. For example, the status of garage lights. This is currently a URI-based scheme (to be documented in more detail within the WIP Attribute Catalog), with the former attribute described as `/garage/lights/status`. 
 
-## Build & Run Locally
+**observation:** a specific measurement taken by one vendor device for a specific attribute. For example, a smart lightbulb chanigng from "off" to "on." 
+
+**produce:** when a vendor device adds an observation to Supertype's data network.
+
+**consume:** when a vendor device gets an observation from Supetype's data network.
+
+## What is Supertype?
+
+Supertype is a data network that allows smart home device makers to harness their ambient data that hubs like Google Home or Alexa aren’t designed to capture, and dev tools like IFTTT or Wink make the end user pay for. Our goal is to create a truly ambient smart home. Any individual can update their smart thermostat, regardless of which brand, and that action is immediately and automatically propagated to all other smart devices in the home.
+
+## How Supertype works
+
+Supertype is a managed data transfer network, with a growing number of touchpoints for vendors of all types to interact with our system. With simplicity in mind, Supertype takes form as an API client library with simple onramps to our data network. Vendors can produce data to and consume data from the Supertype data network.
+
+Unlike other companies, Supertype does not make assumptions on behalf of our vendors. We do not focus routines or predefined actions, our focus lies in getting as much data as possible into vendors' hands.
+
+## Vendor and user login flow
+
+![Vendor Login](internal/images/vendor-login.png?raw=true "Vendor Login")
+![User Login](internal/images/user-login.png?raw=true "User Login")
+
+### Current Support
+
+Supertype curently supports the following onramps:
+- HTTP
+- WebSocket
+
+Supertype currently supports the following API client libraries:
+- Golang
+
+## Encryption
+
+Supertype uses a custom encryption protocol to guarantee efficient end-to-end encryption across all vendor devices. Loosely inspired by the [JEDI Encryption Protocol](https://arxiv.org/abs/1905.13369), Supertype's protocol aims to reduce the number of built-in features with the benefit of only relying on AES encryption, one of the most common forms of symmetric encryption.
+
+The high-level overview of Supertype's encryption protocol is as follows:
+
+1. When a user logs into Supertype through a vendor device, the user's information along with their unique AES encryption key is returned to the vendor.
+2. The vendor can then use this AES encryption key to uniquely encrypt data for this user before uploading it to the Supertype data network. 
+
+A more detailed overview can be found below:
+
+![Encryption](internal/images/encryption.png?raw=true "Encryption")
+
+## Branching Rules
+
+1. All pull requests must contain 1 squashed commit, rebased off the current master branch
+2. Branches and commit messages should have the following prefixes:
+
+    * FEAT: For new functionality
+    * BUG: For bug fixes
+    * DEBT: For cleaning up tech debt
+    * HOTFIX: For a small, ticket-less, direct-to-master hotfix
+
+## Local setup
+###  Prerequisites
+1. Ensure account created on AWS and AWS is configured locally using AWS CLI
+2. Golang, Redis, git installed on local machine 
+
+### Build and Run
 1. `git clone https://github.com/super-type/supertype`
 2. `cd supertype`
 3. `make run`
