--- docs/README.md
@@ -1,3 +1,0 @@
-# `/docs`
-
-Design and user documents (in addition to your godoc generated documentation)
