--- internal/README.md
@@ -1,3 +1,0 @@
-# `/internal`
-
-Where our encryption tooling is contained
