+++ pkg/caching/service.go
@@ -8,13 +8,13 @@ import (
 
 // Repository provides access to relevant storage
 type repository interface {
-	Subscribe(ctx context.Context, conn *websocket.Conn, channel string)
+	Subscribe(ctx context.Context, conn *websocket.Conn, channel string) (*[]byte, error)
 	Publish(ctx context.Context, channel string, messages interface{})
 }
 
 // Service provides consuming operations
 type Service interface {
-	Subscribe(ctx context.Context, conn *websocket.Conn, channel string)
+	Subscribe(ctx context.Context, conn *websocket.Conn, channel string) (*[]byte, error)
 	Publish(ctx context.Context, channel string, messages interface{})
 }
 
@@ -28,11 +28,15 @@ func NewService(r repository) Service {
 }
 
 // Subscribe adds specified attributes to relevant Redis lists
-func (s *service) Subscribe(ctx context.Context, conn *websocket.Conn, channel string) {
-	s.r.Subscribe(ctx, conn, channel)
+func (s *service) Subscribe(ctx context.Context, conn *websocket.Conn, channel string) (*[]byte, error) {
+	message, err := s.r.Subscribe(ctx, conn, channel)
+	if err != nil {
+		return nil, err
+	}
+	return message, nil
 }
 
-// Publish publishes the given message to all subscribers
+// Publish publishes the given message to all consumers
 func (s *service) Publish(ctx context.Context, channel string, messages interface{}) {
 	s.r.Publish(ctx, channel, messages)
 }
