+++ pkg/http/rest/handler.go
@@ -5,6 +5,7 @@ import (
 	"net/http"
 
 	"github.com/gorilla/mux"
+	"github.com/super-type/supertype/internal/utils"
 	"github.com/super-type/supertype/pkg/authenticating"
 	"github.com/super-type/supertype/pkg/caching"
 	"github.com/super-type/supertype/pkg/consuming"
@@ -24,7 +25,7 @@ func Router(a authenticating.Service, p producing.Service, c consuming.Service, 
 	router.HandleFunc("/createUser", createUser(a)).Methods("POST", "OPTIONS")
 	router.HandleFunc("/produce", produce(p, cache)).Methods("POST", "OPTIONS")
 	router.HandleFunc("/consume", consume(c)).Methods("POST", "OPTIONS")
-	router.HandleFunc("/listObservations", httpUtil.IsAuthorized(listObservations(d))).Methods("GET", "OPTIONS")
+	router.HandleFunc("/listObservations", utils.IsAuthorized(listObservations(d))).Methods("GET", "OPTIONS")
 	return router
 }
