--- pkg/producing/connections.go
@@ -1,7 +1,0 @@
-package producing
-
-// MetadataResponse is the vendor metadata used to sync vendors on data production
-type MetadataResponse struct {
-	VendorConnections []string `json:"connections"`
-	Vendors           []string `json:"vendors"`
-}
