+++ pkg/storage/dynamo/consumingImpl.go
