+++ pkg/storage/dynamo/producingImpl.go
