--- pkg/storage/redis/repository.go
@@ -1,6 +1,0 @@
-package redis
-
-// Storage keeps data in dynamo
-type Storage struct {
-	Observation
-}
